+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.67.5] - 2026-04-16
 
 ### Fixed